refactor to socket service but keep same functionalty
